requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
groq
